Cambiato tipo valore calorie

+++ src/main/java/nestorcicardini/D1/entities/Merchandise.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @ToString
 public class Merchandise extends Product {
 
-	public Merchandise(String name, double price, int calories) {
+	public Merchandise(String name, double price, double calories) {
 		super(name, price, calories);
 	}
